refactor: update check exception for delete system (#6)
* feat(system): add CRUD operations for system entity

* feat(upload-file): add function upload file for task

* feat: Add attribute to SystemDTO

* fix: code review
- rename toEpochSeconds toEpochMilli in DateTimeUtil
- refactor LocalDatetime type to Long type
in SystemDTO
- refactor query in SystemRepository
- remove upload-file function

* refactor: update sort for system

* refactor: update check exception for delete system

* refactor: update check exception for delete system

* refactor: update authorize for SystemController
- update in exception handling controller
- update try catch in AuthenticationController
- update in AuthEntryPointJwt

+++ src/main/java/com/sharp/vn/its/management/service/SystemService.java
@@ -1,20 +1,24 @@
 package com.sharp.vn.its.management.service;
 
+import com.sharp.vn.its.management.constants.SortType;
 import com.sharp.vn.its.management.dto.system.SystemDTO;
 import com.sharp.vn.its.management.entity.SystemEntity;
 import com.sharp.vn.its.management.entity.UserEntity;
 import com.sharp.vn.its.management.exception.DataValidationException;
 import com.sharp.vn.its.management.exception.ObjectNotFoundException;
+import com.sharp.vn.its.management.filter.SortCriteria;
 import com.sharp.vn.its.management.repositories.SystemRepository;
+import com.sharp.vn.its.management.repositories.TaskRepository;
 import com.sharp.vn.its.management.repositories.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * The type System service.
@@ -40,6 +44,9 @@ public class SystemService extends BaseService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private TaskRepository taskRepository;
+
     /**
      * Gets all systems data.
      *
@@ -56,10 +63,13 @@ public class SystemService extends BaseService {
     /**
      * Gets all systems data by search param and pageable.
      *
+     * @param request the request
      * @return the all systems data
      */
-    public Page<SystemDTO> loadAllSystemData(String searchParam, Pageable pageable) {
-        Page<SystemEntity> systemEntities = systemRepository.findBySystemNameContains(searchParam, pageable);
+    public Page<SystemDTO> loadAllSystemData(SystemDTO request) {
+        Map<String, SortCriteria> sort = request.getFilter().getSort();
+        buildSortCondition(sort);
+        Page<SystemEntity> systemEntities = systemRepository.findBySystemNameContains(request.getFilter().getSearchKeyword(), request.getFilter().getPageable());
         return systemEntities.map(SystemDTO::new);
     }
 
@@ -84,11 +94,15 @@ public class SystemService extends BaseService {
     /**
      * Delete system.
      *
+     *
      * @param id the id
      */
     public void deleteSystem(Long id) {
         if (id == null) {
             throw new DataValidationException("System id not found");
+        }
+        if(taskRepository.existsBySystemId(id)){
+           throw new DataValidationException("Cannot delete system");
         }
         systemRepository.deleteById(id);
         log.info("System with id {} deleted successfully.", id);
@@ -125,5 +139,36 @@ public class SystemService extends BaseService {
         systemEntity = systemRepository.save(systemEntity);
         log.info("System saved successfully.");
         return new SystemDTO(systemEntity);
+    }
+
+
+    /**
+     * Build sort condition.
+     *
+     * @param sort the sort
+     */
+    private void buildSortCondition(Map<String, SortCriteria> sort) {
+        if (sort.isEmpty()) {
+            sort.put("updatedDate", new SortCriteria("updatedDate", SortType.DESC.getText()));
+            return;
+        }
+        sort.forEach((key, criteria) -> {
+            switch (key) {
+                case "systemId":
+                    criteria.setFieldName("id");
+                    break;
+                case "createdDate":
+                    criteria.setFieldName("createdDate");
+                    break;
+                case "updatedDate":
+                    criteria.setFieldName("updatedDate");
+                    break;
+                case "updateBy":
+                    criteria.setFieldName("updatedBy");
+                    break;
+                default:
+                    break;
+            }
+        });
     }
 }
